+++ addOns/scripts/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Support for Automtion Framework loaddir action, which loads all of the scripts under the specified directory.
 
+### Changed
+- File parameter to `source`, `file` will still work.
 
 ## [45.3.0] - 2024-05-07
 ### Added